+++ worldturtle/ChangeLog.md
@@ -3,10 +3,12 @@
 ## v0.1.1
 
 * Added `sleep` function.
-* Fixed bugs in circle rendering when going in a clockwise direction.
+* Added `rotationSpeed` function.
 * Added the `shiftHue` color function.
+* Fixed bugs in circle rendering when going in a clockwise direction.
 * Fixed the bounds and potential recursion pitfall in internal normalization
   functions.
+* If rotation left 270 degrees will not turn right 90 degrees and vice versa.
 
 ## v0.1.0